+++ app/src/main/java/com/uzlov/moviefind/ui/HomeFragment.kt
@@ -1,10 +1,11 @@
-package com.uzlov.moviefind
+package com.uzlov.moviefind.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.uzlov.moviefind.R
 
 class FirstFragment : Fragment() {
